Move library source out of lib and into the root
This helps me use plugins like the nexus plugin, which wants to be
applied to the root.
