+++ settings.gradle.kts
@@ -7,11 +7,10 @@
  */
 
 plugins {
-    // Apply the foojay-resolver plugin to allow automatic download of JDKs
-    id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+  // Apply the foojay-resolver plugin to allow automatic download of JDKs
+  id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
 }
 
 rootProject.name = "arc"
-include("lib")
 
 enableFeaturePreview("STABLE_CONFIGURATION_CACHE")
