+++ src/main/java/eu/aylett/arc/internal/Element.java
